refactor(app): extract sendFollowupDraft into send-followup-draft.ts
Pure extraction, no behavior change. Moves the FollowupSendInput type,
the draftImages helper, and the sendFollowupDraft function out of
prompt-input/submit.ts into a new send-followup-draft.ts. Context hooks
(useSDK/useGlobalSync/useSync) are referenced via type-only imports.

submit.ts and use-session-followups.ts now import sendFollowupDraft from
the new module; the direct unit coverage moves to send-followup-draft.test.ts.
