+++ packages/app/src/pages/session/use-session-followups.ts
@@ -2,7 +2,7 @@ import { createEffect, createMemo, createSignal } from "solid-js"
 import { createStore } from "solid-js/store"
 import { useMutation } from "@tanstack/solid-query"
 import { followupCommandText, type FollowupDraft } from "@/components/prompt-input/followup-draft"
-import { sendFollowupDraft } from "@/components/prompt-input/submit"
+import { sendFollowupDraft } from "@/components/prompt-input/send-followup-draft"
 import type { useGlobalSync } from "@/context/global-sync"
 import type { useSDK } from "@/context/sdk"
 import type { useSettings } from "@/context/settings"
